requirements.txt
alembic==1.5.4
aniso8601==9.0.1
attrs==21.2.0
autopep8==1.5.5
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Login==0.5.0
Flask-Migrate==2.6.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.4.4
gunicorn==20.1.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.3.2
Mako==1.1.4
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pycparser==2.21
PyJWT==2.3.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.3
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1